src/bot/requirements.txt
aiogram~=2.21
tortoise-orm[asyncmy]~=0.19.2
cryptography